// apps/web/src/types/interface.ts
export type ReportProgramRecommendItem = {
  title: string;
  cta: string;
};

export type ReportProgramRecommend = {
  challengeCareerStart?: ReportProgramRecommendItem;
  challengePortfolio?: ReportProgramRecommendItem;
  challengePersonalStatement?: ReportProgramRecommendItem;
  live?: ReportProgramRecommendItem;
  vod?: ReportProgramRecommendItem;
  reportResume?: ReportProgramRecommendItem;
  reportPersonalStatement?: ReportProgramRecommendItem;
  reportPortfolio?: ReportProgramRecommendItem;
};